[FIX] move modal
modal.js의 위치 이동
